# backend/openapi.json
{
  "components" : {
    "schemas" : {
...
      "CreateFlagRequest" : {
        "properties" : {
          "description" : {
            "type" : [ "string", "null" ]
          },
          "enabled" : {
            "type" : "boolean"
          },
          "key" : {
            "pattern" : "^[a-z][a-z0-9-]{1,49}$",
            "type" : "string"
          }
        },
        "required" : [ "key" ],
        "type" : "object"
      },
...
      "Flag" : {
        "properties" : {
          "createdAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "createdBy" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "enabled" : {
            "type" : "boolean"
          },
          "fullKey" : {
            "type" : "string"
          },
          "groupId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "key" : {
            "type" : "string"
          },
          "updatedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "updatedBy" : {
            "type" : "string"
          },
          "version" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "required" : [ "createdAt", "createdBy", "enabled", "fullKey", "groupId", "id", "key", "updatedAt", "updatedBy", "version" ],
        "type" : "object"
      },
...
      "ToggleFlagRequest" : {
        "properties" : {
          "enabled" : {
            "type" : "boolean"
          }
        },
        "required" : [ "enabled" ],
        "type" : "object"
      },
      "UpdateFlagRequest" : {
        "properties" : {
          "description" : {
            "type" : [ "string", "null" ]
          },
          "enabled" : {
            "type" : "boolean"
          },
          "version" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "required" : [ "version" ],
        "type" : "object"
      },
...
    }
  },
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/v1/admin/flags/{flagId}" : {
      "delete" : {
        "operationId" : "delete_1",
        "parameters" : [ {
          "in" : "path",
          "name" : "flagId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        },
        "tags" : [ "flag-controller" ]
      },
      "patch" : {
        "operationId" : "update_1",
        "parameters" : [ {
          "in" : "path",
          "name" : "flagId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdateFlagRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Flag"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "flag-controller" ]
      }
    },
    "/api/v1/admin/flags/{flagId}/toggle" : {
      "post" : {
        "operationId" : "toggle",
        "parameters" : [ {
          "in" : "path",
          "name" : "flagId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ToggleFlagRequest"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Flag"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "flag-controller" ]
      }
    },
...
    "/api/v1/admin/groups/{groupId}/flags" : {
      "post" : {
        "operationId" : "create_1",
        "parameters" : [ {
          "in" : "path",
          "name" : "groupId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateFlagRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Flag"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "flag-controller" ]
      }
    },
...
  },
  "servers" : [ {
    "description" : "Generated server url",
    "url" : "http://localhost"
  } ]
}